Fixed issue (#19) where Paver does not render in IE/Edge
IE/Edge supports the DeviceOrientationEvent but does not feed any gyroscopic data: this browser is left out when performing a setTimeout fallback check if gyroscopic data is available. Removing user agent detection fixes this issue, and does not have an ill effect on devices with gyroscopes.

Also updated the code to use guard clauses and replacing undefined with void 0 checks.